requirements.txt
addict==2.4.0
ansi2html==1.8.0
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
easydict==1.10
executing==1.2.0
fastjsonschema==2.18.0
filterpy==1.4.5
Flask==2.2.5
fonttools==4.42.1
freetype-py==2.4.0
hsluv==5.0.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.24.4
open3d==0.17.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.3
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tenacity==8.2.3
threadpoolctl==3.2.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vispy==0.14.0
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.9
zipp==3.17.0
